requirements.txt
# Web Framework
streamlit>=1.30.0

# Data Processing
pandas>=2.0.0

# Database Connectivity
mysql-connector-python>=8.0.0

# Utilities
typing-extensions